+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - YYYY-MM-DD
 ### Added
+- /health endpoint [@wujuu]
 ### Changed
 ### Fixed
 - engine_version in the recommendations response [@Michal-Kolomanski]